Power BI multiple table support/Power BI/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
pydantic==2.4.2
python-multipart==0.0.6
google-generativeai>=0.3.2
pyodbc>=5.2.0 